knative_functions/python/get-emojis/requirements.txt
parliament-functions==0.1.0
emoji==2.12.1
requests==2.32.3
